// types/database.bingo.ts
export interface BingoCardsTable {
  Row: {
    created_at: string | null
    creator_id: string | null
    description: string | null
    difficulty: DifficultyLevel
    game_type: GameCategory
    id: string
    is_public: boolean | null
    tags: string[] | null
    title: string
    updated_at: string | null
    votes: number | null
  }
  Insert: {
    created_at?: string | null
    creator_id?: string | null
    description?: string | null
    difficulty: DifficultyLevel
    game_type: GameCategory
    id?: string
    is_public?: boolean | null
    tags?: string[] | null
    title: string
    updated_at?: string | null
    votes?: number | null
  }
  Update: {
    created_at?: string | null
    creator_id?: string | null
    description?: string | null
    difficulty?: DifficultyLevel
    game_type?: GameCategory
    id?: string
    is_public?: boolean | null
    tags?: string[] | null
    title?: string
    updated_at?: string | null
    votes?: number | null
  }
  Relationships: [
    {
      foreignKeyName: "bingo_cards_creator_id_fkey"
      columns: ["creator_id"]
      isOneToOne: false
      referencedRelation: "users"
      referencedColumns: ["id"]
    },
  ]
}